go.mod: fix up go-ipfs fiasco with golangci-lint
They separated test dependencies from the main module in mid-September,
so update to that version that is just nine days newer than the one we
used before. This allows us to not require golangci-lint at all, so we
can remove all the replace workarounds.

While at it, change the go version to target the current stable 1.13.