test: remove remaining sleeps
The first sleep didn't appear to have a point. Its previous statements
don't start a server, fire a goroutine, or otherwise do anything that
could warrant a sleep.

The second sleep is trickier, as it happens for every request sent. The
sleep appears right before we read a response to a request. However, as
far as I can tell, it appears unnecessary; ReadMessage on the websocket
should block until a message is ready to be read.

Running the test ten times seems to confirm that neither of those sleeps
was necessary, at least under normal circumstances.

While at it, stop ignoring ReadMessage errors.

Reduces 'go test' time from ~26s to ~10s.