all: various code simplifications
Replace Open + ReadAll with ReadFile.

Rename goneturl to neturl; we are in Go code after all.

Rewrite a func lookup if/else chain to be shorter and less indented.

Remove broken checks for overflows; 2^64 is a logical XOR. We could use
2<<64 for the actual limit of uint64, but an uint64 already wraps back
to 0 on its own, so simply rely on that.

Finally, remove some redundant types, and rename "vlog" to just "log"
for consistency with all the other packages.