rewrite net.WebsocketHandle to handle many clients
The previous implementation read messages from connected clients in a
synchronous round-robin fashion. This meant that if A and B both connect
in that order, we would not read a message from B until we've read a
message from A.

Instead, read messages from each client in a separate goroutine, as they
connect. Use an internal channel and goroutine to then route all of
those messages to the receiver.

This also means we don't need the client pool with its channels and map;
each connected client is simply one more goroutine, since HTTP handlers
are called in new goroutines anyway.