Implement authomatic census import (aka census Oracle) on Gateway
Read ethereum logs and for each "process created" event, take the census
merkle tree information, download and import it into the local storage.
The censusId will be equal to the merkle root.
As the only storage supported for now is IPFS, implements a non-blocking
queue mechanism.

Using this feature with ipfsSync makes the census synchronitzation
between gateways faster and more efficient.