cmd/gateway: enforce a minimum of open files
The websocket handler did attempt something very similar, but it obeyed
the hard limit without question. If the hard limit is too low, such as
1024, we should refuse to start. Otherwise, the gateway would likely
crash a bit later when reaching the low limit.

Doing it in cmd/gateway is more natural, and it also lets us notify the
user of this global process setting being changed:

	INFO gateway/gateway.go:543 raising file descriptor soft limit from 1024 to 524288

We can also see that the required minimum makes the gateway refuse to
start:

	$ ulimit -Hn 1024; go run ./cmd/gateway
	[...]
	FATAL gateway/gateway.go:220 could not ensure we can have enough open files: hard limit is 1024, but we require a minimum of 8000

The requirement of 8000 is chosen arbitrarily, as it's a value roughly
eight times larger than the 1024 that has been giving us trouble.

Most Linux boxes would allow for numbers as high as 50k or 500k, but
unless we absolutely need them, it's best to stick to lower figures for
now. This will keep the gateway more portable and less resource
intensive.