CI: simplify the setup, now centered around Go
Stop using docker-in-docker for everything. The main purpose of CI,
which is to run the tests, can be done directly with a Go image.

Start using the cache too, so that we don't need to re-download and
re-build the same stuff at every commit and PR. Both of these steps
forward allow us to greatly cut jobs from ~12m to ~1m30s.

While at it, make it so that PRs don't need to do any docker login or
build. Those are only needed for the step that releases a Docker image.

Finally, don't set global defaults for the image and services. Those
only apply to the docker-release step; starting the dind service in
other steps is wasteful.