Implement process cancelation and process lifecycle:
Now a process can be canceled in ethereum and the oracle will send a
tx to the vochain to cancel the process living in the voting chain.

Changing the processes internal state cannot be done automatically without
a txs that indicates to do so.
An oracle should be sending tx for each process and monitorizing
the blocks too often.
To face this, a process now have a field called Canceled and all the
process lifecycle logic will be handled using this boolean and the
network chain height. The oracle will only send a tx if the process is canceled.
A new vote will be added to the process only if the process is not canceled
and is sended during the valid block frame.

- oracle handles ProcessCanceled ethereum event
- delete current process state strings, use Canceled boolean instead
- iavl can handle a process cancelation
- added vochain logic for supporting process cancelation
- added vochain logic to not allow votes if process is canceled or
  the blockframe is not correct

TBD: adapt the scrutinizer since currentProcessState no longer exists