census: fix a major data race in the manager
The manager had a mutex to attempt to protect itself from data races,
but it was used only in two methods. A dozen other pieces of code read
from the Trees map without taking care of the mutex.

First, make AddNamespace return the new merkle tree, so that the pieces
of code that need to add a namespace and then interact with the tree
don't need to access the Trees map directly.

Second, add the proper locking to the remaining bits of code that still
need to access the Trees map.

With this change, TestBenchmark is able to pass a few runs with -race.