census: make AddToImportQueue non-racy
First, unexport ImportQueue and ImportQueueDaemon, since neither of them
should be accessible from other packages.

Second, replace the map with a channel. A map isn't safe for concurrent
use, and it was simply being used as a queue - a channel is perfect for
that. On top of fixing the data race, we can simplify the code by
removing the double for-loop, map deletes, and the 1s sleep.

This also has the big benefit that, when AddToImportQueue is called, the
census import will happen instantly if the queue was empty. Before, it
could take up to 1s for the "daemon" goroutine to come out of sleep and
do the work, even if the queue was empty.

While at it, flip the condition on len(dump.ClaimsData), so that we can
further unindent the last dozen lines of code.