all: start enforcing staticcheck
Fix all the remaining issues, and add it to CI.

Use a separate parallel step, so that we can run tests and static
analysis in parallel. They only duplicate a bit of work to download the
modules and compile dependencies, but that's cached anyway.

Finally, to run staticcheck, download an official prebuilt binary.
Building it ourselves would be much slower, given how Gitlab's public
runners are pretty slow.