router: refactor buildReply to deduplicate code
Practically all the endpoints did all of these steps:

1) Set up a few basic response fields, like the ID and timestamp.

2) Sign the MetaResponse contents (JSON-encoded)

3) Encode the whole response in JSON

Do all of this as part of a Router.buildRequest method, so we can remove
over two hundred lines of copy-pasted code. This way we also ensure that
any endpoint returns the same kind of non-error responses.

While at it, simplify sendError a bit.

Fixes #103.