router: simplify the method map setup
First, use a single map instead of two. This makes registering methods
and looking them up simpler, and avoids the edge case of a method name
being both a private and a public method.

Second, reject duplicate methods.

Third, use method values, so that all the implementations can be methods
on Router directly. This also keeps the "handler" function type simpler,
as the request input is one parameter.