all: try to fix data races spotted in some tests
The tests use log.InitLogger pretty offten. This is racy, because the
function sets up a global logger, and that can only be done once and
before any tests run.

Instead, use TestMain for that purpose in both the test and benchmark
packages, and remove InitLogger calls elsewhere.

While at it, consolidate rand.Init calls, remove an unnecessary use of
rand, simplify the signature of DvoteAPIServer.Start, and fix the uses
of "debug" logging in some tests.