census: compress census dumps on ipfs
From the logs on TestCensus, we can see some realistic-looking stats:

	compressed 52.31 KiB to 4.82 KiB in 1.935794ms with zstd, 9.2% of the original size
	[...]
	decompressed 4.82 KiB to 52.31 KiB in 169.661µs with zstd, 1084.9% of the original size

That is, we reduce space by 10x, by using milliseconds to compress, and
hundreds of microseconds to decompress. This is on my 2017 ultrabook.

We picked zstd since it provides compression ratios similar to gzip,
while being multiple times faster at compressing and decompressing.
Alternatives like lz4 would also be fast, but not compress very well.

A big advantage of zstd is that it's becoming the new standard when one
has a tradeoff between compression ratio and speed. Another big point
here is that zstd has a pure Go implementation which hit v1 last year,
which is what we use here.

The implementation is decoupled into its own type, so that we can unit
test it to ensure its properties. It's also tested as part of
integration tests like TestCensus. The decoupling also allows us to
reuse this in other components in the future.

Moreover, the decompressBytes method detects the format of the
compressed data. This way, we can support importing non-compressed
census dumps too, such as those generated by older gateways or clients.

Fixes #168.