test: re-enable eth_test, making it non-racy
Unfortunately, ethereum's node.Stop is racy, so we can't use it. See the
linked issue in the TODO.

While at it, make HTTPPort and WsPort random, so that we can run the
test twice in a row without failures (assuming we use the racy
node.Stop). Add a TODO about a slightly better approach in the future.