stub out the duktape Go dependency entirely
It's useless, and adds a huge bottleneck to our build times. The numbers
below are on my laptop from incremental builds, where we have to rebuild
duktape.

Before:

	$ time go build ./cmd/gateway
	# gopkg.in/olebedev/go-duktape.v3
	In file included from _cgo_export.c:4:
	debugger.go:22:13: warning: ‘_duk_debugger_attach’ defined but not used [-Wunused-function]
	# gopkg.in/olebedev/go-duktape.v3
	duk_logging.c: In function ‘duk__logger_prototype_log_shared’:
	[warnings]

	real	0m31.509s
	user	0m36.112s
	sys	0m2.138s

After:

	$ time go build ./cmd/gateway

	real	0m8.254s
	user	0m11.992s
	sys	0m1.341s

The binaries also get a bit smaller, going from ~86MiB to ~84MiB
unstripped:

	$ du gateway*
	86560	gateway_before
	84168	gateway_after

Since this is a replace directive in go.mod, it will affect any
module-mode builds run within this module, which is fine for us. This
way, it will speed up Docker builds too.

The package's API hasn't changed in multiple years, so it's unlikely
we'll have to update the stub anytime soon. And if we have to, that's
okay.