remove the use of github.com/pkg/errors
Its only purpose was to support error wrapping. Since Go 1.13, the
standard library now supports it via fmt.Errorf and the "%w" verb, so
use that instead.

While at it, lower the timeout for 'go test' on CI. The timeout is for
each package separately, so one minute is plenty. Our slowest package is
./test, which takes 2-4s to test. The default timeout of 10m meant that,
if anything got stuck, CI would waste a lot of time.