docker: ensure images include ca-certificates
In 74ecbe1f4e I refactored our dockerfile to deduplicate code, and
produce smaller images. This was fine for the most part, but we forgot
about one crucial piece.

The debian "slim" images are truly slim, so they don't contain
ca-certificates by default. Add it. Without it, we'd get errors like:

	Get "https://acme-v02.api.letsencrypt.org/directory": x509: certificate signed by unknown authority

Reuse the file from the golang debian image, instead of using apt to
install the package, because a single file copy is significantly faster
and doesn't require network access at every incremental rebuild.

Proof that the file is now included, where 'gateway1' and 'gateway2' are
the before and after docker images:

	$ docker run --entrypoint=/bin/sh gateway1 -c 'find /etc/ssl'
	find: '/etc/ssl': No such file or directory
	$ docker run --entrypoint=/bin/sh gateway2 -c 'find /etc/ssl'
	/etc/ssl
	/etc/ssl/certs
	/etc/ssl/certs/ca-certificates.crt