Revert "docker: don't download unnecessary modules"
This reverts commit 9f9a1adadee544e622990b82f4f6f942eac27a77.

Reason for revert: 'go list ./...' matches no packages at that stage,
since we only added go.mod and go.sum and not the Go source code. So the
step no longer downloads and caches any dependencies.

Doing the 'COPY . .' sooner would be problematic as well, as we'd then
need to re-download all the dependencies anytime a single line of source
code changes.

'go mod download' is the best we can do, right now. Revert the change.