subpub: fix a channel race bug, improve the test
The bug was in the peerChan loop in SubPub.discover. When we iterated on
a peer that we weren't interested in, such as ourselves or a peer we
already connected with, we would break instead of continue the loop.

This resulted in a channel race. For example, if our own node showed up
before other nodes, we'd ignore the other node entirely. This resulted
in the test getting stuck about 10% of the time.

Another real bug was that SubPub.advertise wasn't checking errors
properly, so the test was setting the sleep duration to 0s. This is
equivalent to a cpu-busy loop, so the test consumed a lot of resources.
Check the error, and ensure we're sleeping at least 1s between
iterations to avoid consuming lots of cpu.

The test is now more resilient, too. For example, it uses random ports,
it uses a timeout for all channel send/receive operations that might
block forever, and it cleans up after itself better with defers.

While at it, make the API a bit more context-aware, which helped make
the test better.