all: update tendermint dependencies
This change updates all tendermint modules to their latest releases.

One blocker that prevented us from doing this in the past was that tm-db
depended on bbolt incorrectly, resulting in failures from 'go get'. We
work around this here with a temporary replace directive. Upstream will
presumably fix this in the upcoming v0.5.1.

Aside from the above, there are a bunch of backwards incompatible
changes we had to accomodate to. They were mainly minor ones, such as
renaming or moving names, or adding extra parameters to functions, or
making functions return errors instead of panicking.