subpub: refactor to fix two bugs, improve the test
First, improve the test so that we have one "me" node, plus five
"remote" peers to connect to. This is far more realistic than just
talking to a single remote peer.

Second, fix a bug unearthed by the change above. Broadcast messages were
being sent to a single random peer, instead of all peers. This was
because a single channel was shared by all peer writer goroutines -
thus, only one would grab the broadcast message, and the others wouldn't
get it at all.

To fix this issue, add another channel and goroutine, which "fans out"
the message to each of the broadcast channels for each goroutine. This
required a bit of a refactor, as we now need extra channels.

Third, fix a previously encountered bug - discarding disconnected peers
would easily panic. This wasn't being caught before, since we only had
one peer and it never disconnected. The test now disconnects two peers,
which reproduces the crash, and the crash is now fixed. The loop now has
a comment to explain why we can't use a range statement.

Fourth and last, the use of context cancellation on libp2p along with
its Close methods is racy. Consistently use the Close methods alone,
which can block and return errors, and which are the preferred way by
upstream.

Fixes #200.