all: use full URL for tendermint/rpc/client.NewHTTP
That is, "tcp://host:port" instead of just "host:port". The latter is
often considered an address, such as when doing a TCP listen, but it's
not a valid URL given the lack of a scheme.

The reason this issue surfaced is because NewHTTP started returning an
error. It's unclear why it worked before with an invalid URL;
presumably, it allowed our old kind of address, or silently ignored the
error.

Replace some uses of fmt.Sprintf and strings.Split for joining and
splitting "host:port" addresses, too. Their counterparts from the proper
net package have proper IPv6 support, and will give good errors.

While at it, clean up some direct uses of the log package in tests.

Fixes #201.