all: unify benchmarks again, test cleanups
There isn't a good reason to use one package per benchmark. The only
reason we were forced to do this is because the flags were modified
in-place. Stop doing that, and then they can easily be shared. This also
allows us to re-run the tests, since the flag values don't change after
the first run.

The testcommon functions take a testing.TB for errors, so they will
never return a nil pointer. Remove all of those checks.

In NewMockVochainNode, stop the node when the test or benchmark
finishes.