all: switch to Go 1.14.2
This should fix the runtime self-deadlock issue we've been seeing.

While at it, double the test timeout with -race. We were seeing some
test timeouts with -race on CI. Here are some measurements from my
laptop:

	$ go test ./test
	ok  	gitlab.com/vocdoni/go-dvote/test	3.674s
	$ go test -race ./test
	ok  	gitlab.com/vocdoni/go-dvote/test	49.068s

That is, -race easily makes the tests run 10x slower. Since we're using
1m as a timeout for the regular test run, let's use 10m for the one with
the race detector.

Ten minutes is quite high, but I don't think we've got much of an option
here. The race detector is useful, and the public CI runners are pretty
slow.