introduce --minerKey flag
Instead of using a disk file for persistent tendermint mining key
storage, use a flag --minerKey=<hexString 64 digits>

If the flag is specified, use the provided key for block signature. If
flag not specified, generate new one or load an existing one.

When --minerKey is used the private key is automatically saved to disk.
So there is no need to specify on each run the key.

ENV files and docker config files have been adapted but not tested.