simplify error logging to a file
Doing this for the "fatal" level is pointless, because a log.Fatal calls
os.Exit, as documented by the zap library.

Delay spawning the goroutine until we actually need to write. Otherwise
we're spawning goroutines that do nothing if the feature is disabled.

Make the function return an error instead of panicking, while at it.

We can also use log.Fatal instead of panic when loading the
configuration initially, as the log package is initialized with default
settings on its own.