Fix IterateRange on iavl The IterateRange method receives an start and end. If we only put start = "<processID>+`_`" and end = nil the mentioned method returns all values from the start key until the end of the tree so we get undesired processID and therefore undesired vote nullifiers.
In order to solve that, given that processID and Nullifier are hashes,
fixed length strings, and given that we store envelopes as "pid_nullifier"
we can stay sure that the char `_` will always stay in the same position,
so it is fair to add as the end value in IterateRange method the char `}`.
So now we have start = "<processID>+`_`" and end = "<processID>+`}`".

We can simplify countVotes iavl method that ran well luckily because
we were getting all leafs and filtering with an if that returns true
if the prefix is the processID.

And we also can extract the IteratorRange function to be usable anywhere.