crypto/signature: simplify API around Address
First, remove StringFromAddres, since it was entirely unused.

Second, don't expose AddressFromString, since it was only really needed
internally. The vochain package should be using AddAuthKey.

Third, remove Address.String, as it's not necessary. Go already
pretty-prints arrays with hex, and we were relying on that behavior for
the tests. Adding our own String method into the mix was confusing.