crypto/signature: avoid hex roundtrip in Encrypt/Decrypt
We were roundtripping to hex to convert from *ecdsa.PublicKey to
*secp256k1.PublicKey, and the same situation applied to private keys.

However, the secp256k1 type is just a wrapper:

	type PublicKey ecdsa.PublicKey

This means that we can just convert from one type to another, without
having to encode and decode.

We need a separate variable since we need a pointer as a result.