add block timestamp on getBlockHeight api response
iavl height method changed to header in order to return the whole header,
then we can take the height easily. Adapt the rest of the code that previouly
used iavl.Height() to use iavl.Header()

change voteCallbacks getBlockHeight to use the ABCI queries defined,
for now all this kind of requests are using the ABCI query method.
We are not using blockstore anywhere and we are relying on our iavl trees,
so for consistency use iavl methods everywhere.