crypto/signature: don't use pointers inside SignKeys
We already use a pointer to SignKeys in most scenarios, like *SignKeys
or []*SignKeys, which is also the case for all of its method receivers.

Putting the public and private keys behind another pointer will just
lead to double indirection, which will make the code jump around memory
twice for no reason.

Moreover, the two types are tiny; for example, PrivateKey only weighs as
much as four pointers.