census: use a single database in the Manager
Instead of one new database per census (merkle tree), with its own
sub-directory, we reuse the same parent long-running database with a
prefix. This simplifies the tree package, and allows the census manager
to hold hundreds of censuses at once without having to keep hundreds of
databases open.

This is particularly important now that we use the badger database. It's
heavily optimized to hold gigabytes of data and heavy read/write
traffic, so it uses a few goroutines and dozens of megabytes even when
sitting idle. Sharing one database for all censuses scales well, and
doesn't use nearly as much memory as one database per census.

The only downside here is that we no longer support deleting a census
from the database, since the iden3 database layer doesn't have that
feature (iterating over a prefix and deleting those entries). A TODO is
left for that, to be tackled later, since one can always delete the
entire dvote directory to clear all databases.

The following command was run to obtain the benchmark numbers. Since we
want to test how well the manager does at adding many censuses, we use
benchtime=20x, so that 20 censuses are added. Each with a small size of
just 100, as we don't care about the census size all that much here.

	go test -bench=Census -benchtime=20x -cpu 1 -count 5 . -censusSize 100 -onlyCreateCensus

As one can see by the numbers, performance is largely unaffected, but
memory usage decreases by as much as 70%. Our memory usage should no
longer increase as quickly with the number of imported censuses.

The number of allocs/op (objects allocated per census) increases
slightly. The reason behind that isn't clear. Though we care about
alloc/op mainly, as that will be the main driver behind memory usage for
long-running gateways.

	name    old time/op    new time/op    delta
	Census     255ms ± 3%     251ms ± 3%     ~     (p=0.548 n=5+5)

	name    old alloc/op   new alloc/op   delta
	Census    29.9MB ± 1%     8.8MB ± 2%  -70.58%  (p=0.008 n=5+5)

	name    old allocs/op  new allocs/op  delta
	Census      140k ± 2%      145k ± 1%   +3.89%  (p=0.008 n=5+5)

Fixes #258.