avoid using ABCIQuery, use db gets instead move vochain client to BaseApplication move processSize, entityIdSize and nullifierSize into consts in order to reuse them, also add some sanity checks
All calls to the vochain from outside are made through the router
so using RPC on the same machine, under heavy load, can produce errors or
some unexpected behaviours.
We thought that it is better to make that calls directly to the database
which is optimized for doing so. With this approach we lose the Tendermint
Core ability to return merkle proofs of the data fetched.
Currently we are not using this feature, but if at some point we need it
we can generate our merkle proofs without depending on Tendermint and it
should be pretty straight fordward to do so given the codebase we have.