use scrutinizer as work around to getProcessList until vochain refactor is done
With the Vochain refactor, TM events were removed so it is not
possible to get a process list given an entityID.
The idea is to index the entityID in the iavl ProcessTree but this
comes with a lot of changes. So for now let's use the scrutinizer
onProcess event to use the empty value of each entity entry in the databse
to store the process IDs of an entity.
The format is: key: "e_entityID" value: "["pid1", "pid2", ...]"

Also added in this commit the possibility of execute cmd/vochaintest with
a predefined entityID

Finally clean the vochain start events config because is no longer needed for getProcessList