vochain/keykeeper: minor code review improvements
This package uses crypto/nacl, so reuse its KeyLength constant.

A defer in an endless loop never runs, so that's a bug and a memory
leak. Found by staticcheck.

Use early continues with an iterator loop, to prevent extra indentation.

Simplify some variable declarations.

Add TODOs for code that should be improved once we finish the refactor
of the database API.

Finally, use go-cmp in the test, which is simpler and gives better
human-readable results.