use nhooyr ws library instead of gorilla
First, we used the gorilla websocket Close() method that closes the
connection without sending a close message, it just closes the
connection abruptly. If this happens during a ReadMessage() an EOF
error is triggered and handled with a Fatal which makes the app stop
running.
It becomes difficult to handle this situation because the gorilla Writer
sends a Close() message at the end of every read.

Second, currently the gorilla websocket library does not have an official
mantainer and all changes proposed like the use of Go Context or the
implementation of a gracefully shutdown method are frozen during months.
Even is a widely used library we need to depend on libraries that are
actually mantained and updated consequently with the Golang improvements.

So I think it is a good idea to stop depending on an unmantained library.
The use of nhooyr library allows use to easly close connections in a
graceful way and also allows us to make use of the Golang Context.

You can see an in depth comparation at:
https://github.com/nhooyr/websocket/#gorillawebsocket

Third, this commit makes the appropiate changes in the code
in order to use the nhooyr websocket library. Most of the changes
are just function renaming.
The Upgrader method used by gorilla is not required anymore.
The changes are mostly a plug&play replacement.

Fourth, all tests where executed and I cannot see an abnormal
websocket close causing the panic in tests.

Fifth, the websocket lock is no longer needed because was
only used in the Write operation. The new library
already implements a Lock in Write() method so we do not
want two lock levels.

Sixth, the websocket configurable buffer sizes were also deleted. The
reason for doing so is that currently golang/net does not implement
hdr buffers so it makes no sense to have it in an upper layer.
I do not think the buffer sizes will be a near problem, so if
at some point it could be we will see.

Finally, note that this new library can target WASM and may be
useful for the future. Also we can get rid of one write at a time
because nhooyr supports concurrent writes.

Thanks @mvdan for the tips <3