vochain: don't use tendermint/iavl in a racy way
Any read or write operations on a IAVL tree, mutable or immutable, must
be guarded with a mutex. See https://github.com/tendermint/iavl/issues/241.

We were only guarding reads and writes to the immutable tree, but that
isn't enough. A write on a mutable tree can modify its corresponding
immutable tree, for example if the immutable tree needs to be rebalanced
after a key was deleted or added.

As such, guard absolutely every IAVL operation with the same RWMutex.

Fixes #262.