vochain: temporarily recover some iavl panics
We've seen a handful of panics from iavl in Envelope and EnvelopeList
calls.

Until https://github.com/tendermint/iavl/issues/212 is fixed by
upstream, try to return or log an error instead of crashing the entire
gateway.

We shouldn't see these panics anymore after we fixed the data races in
the previous commit. Still, be extra safe.

For #262.