fix iavl ImmutableState
First, when updating ImmutableState on beginBlock() it was done by assigning
the inner immutable tree of each mutable tree to each Immutable
State tree. This is wrong because the correct way of getting
the immutable tree is trough the GetImmutable() method.

Second, now we update the ImmutableState on beginBlock() using the
State.Immutable() method which loads an immutable tree at latest
saved version.

Third, ImmutableState is available when current block is 1 or higher.
This is because we cannot get a version until it is saved
to disk.