vochain: greatly simplify events
First, since the interace forces implementing all the event handling
methods, don't support registering a handler on a per-method basis. A
function call is very cheap, and the KeyKeeper already has empty method
declarations for this anyway.

Second, this means that we can replace the map[string][]T with just []T,
and remove ValidEvents entirely. We also no longer have two names for
each event, which didn't even match - for example, addVote and OnVote.

Third, rename Event to EventListener, and AddEvent to AddEventListener,
for clarity, and fix the docs a bit.

Fourth and last, simplify the code overall; we no longer need loops to
call event listeners, and AddEventListener has no need to return an
error. The error was being ignored before, too.