crypto: redesign crypto interface to remove KeyPair
The old KeyPair is now just PrivateKey, since one can derive the public
key from a private key.

The addition is PublicKey; thanks to the separate type, the following
commit will be able to extend the API to allow Cipher to encrypt data
for an arbitrary public key, and not just the same publick key derived
from the private key doing the encryption.

The key types are also interfaces now instead of []byte directly. This
means that the API can use the underlying values directly (such as byte
arrays) without having to copy from a []byte.

Obtaining a []byte can still be useful to represent a key in plaintext
or binary form, so we add a Bytes method to the interface.