crypto: refactor encryption to take a recipient
This way, we can easily encrypt a message for a specific public key that
doesn't correspond to the public key derived from the private key doing
the encryption. This was exposed in the nacl package via a separate
static function, but we now expose it via the same method.

Since nacl supports encryption without a private key, expose an
Anonymous variable to allow such encryption. This is similar to the old
static function, but it uses the same interface and method.

While at it, split Verify into an interface that only requires a public
key, since a private key is generally not required to verify a signed
message.