dockerfiles: reuse images in docker-compose
Running 'docker build .' takes about ten seconds with the current
version of Docker, even if all the steps are cached. Presumably, this is
because of the extra overhead of setting up the context, computing each
layer, and figuring out that no work needs to happen.

Unfortunately, docker-compose has a total of eight services, and seven
of them all try to rebuild the same image:

	$ time docker-compose build
	Building seed
	Building miner0
	Building miner1
	Building miner2
	Building miner3
	Building oracle
	Building gateway
	Building test

	real	1m14.043s
	user	0m50.874s
	sys	0m3.097s

Instead, since the miners, oracle, and gateway all build the same image
with the same tag as the seed service, and since they all depend on it
to start anyway (see "links: seed"), we can just reuse the image without
building it at all.

	$ time docker-compose build
	miner0 uses an image, skipping
	miner1 uses an image, skipping
	miner2 uses an image, skipping
	miner3 uses an image, skipping
	oracle uses an image, skipping
	gateway uses an image, skipping
	Building seed
	Building test

	real	0m17.015s
	user	0m13.332s
	sys	0m0.754s

Better yet, if buildkit (aka buildx) is available, the caching of layers
is far, far cheaper. Do that in the CI config:

	$ time COMPOSE_DOCKER_CLI_BUILD=1 DOCKER_BUILDKIT=1 docker-compose build
	[...]

	real	0m1.097s
	user	0m0.817s
	sys	0m0.176s

More info below:

	https://www.docker.com/blog/faster-builds-in-compose-thanks-to-buildkit-support/

This helps with local testing, but should also help with CI greatly, as
it saves about a full minute of the CPU spinning for no reason.