extend getEnvelopeStatus api call
First, add Height on Vote struct in order to store the block number
where the vote was saved.

Second, iavl addVote() sets the vote height before the vote is saved.

Third, modify getEnvelopeStatus api call to return the block
where the vote was saved and its corresponding timestamp.
The timestamp is obtained using the reference to the local running node
which allows to get a given block through the node's block store.