Revert "use ethereum startNode function"
This partially reverts commit c1b1c6b8209026f6884921b26ea2637bbd5c5514.

I had removed the use of utils.StartNode in
257707059771ff0315d0adbfa24633fcddf7be08 back in March. Unfortunately, I
only documented why in the commit message, not the code. It's entirely
normal that someone else read the code and tried to simplify it again.

Explain the reason properly in the code, to warn future developers from
trying to roll this back again.