Dockerfile: use a small FROM in all targets
The test target was using "FROM builder", which has the unfortunate side
effect of including all of Go and the module and build caches in the
final image. The test image ended up weighing over 2.5GiB.

After this change, it only weighs 85MiB, which is similar to the main
dvotenode image. This is not hugely important, since we don't push the
test image anywhere, but it still makes docker use far more resources
and disk to build and store it.