all: simplify error comparisons
When declaring global errors, declare them as errors instead of strings.
That way, we can check for them without having to use string
comparisons, which are hacky and bound to break at some point.

This does turn one globals const to a var, since the errors package
doesn't provide a way to declare constant errors. We could do that
ourselves with our own error type, but it's overkill for these simple
errors. We can trust our code to not modify the globals.

The other global constant was copied from the badger repository. Use
their error variable directly for now, with a TODO to abstract that away
in our db package. Don't do that here, as that would be a larger
refactor.